requirements.txt
# 数据采集
feedparser>=6.0.0
requests>=2.31.0
aiohttp>=3.9.0

# AI 处理
anthropic>=0.40.0

# 配置管理
pyyaml>=6.0.0

# HTML 生成
jinja2>=3.1.0

# 工具
python-dateutil>=2.8.0
